+++ src/de/kaffeeshare/server/plugins/BasePlugin.java
@@ -15,15 +15,22 @@
  ******************************************************************************/
 package de.kaffeeshare.server.plugins;
 
+import java.io.IOException;
 import java.net.URL;
 import java.util.logging.Logger;
 
+import org.jsoup.Jsoup;
 import org.jsoup.nodes.Document;
 
+import de.kaffeeshare.server.datastore.DatastoreManager;
 import de.kaffeeshare.server.datastore.Item;
+import de.kaffeeshare.server.exception.SystemErrorException;
 
 /**
  * Abstract base class for all plugins.
+ * 
+ * A plugin has to implement the "match" method to select the urls handled by the plugin.
+ * Specialized outputs are generated by overriding the methods.
  */
 public abstract class BasePlugin {
 
@@ -35,14 +42,111 @@ public abstract class BasePlugin {
 	 * @return true, if plugin match otherwise false
 	 */
 	public abstract boolean match(URL url);
-	
+
 	/**
 	 * Create an item.
 	 * @param url URL
 	 * @return Item
 	 */
-	public abstract Item createItem(URL url);
+	public Item createItem(URL url) {
+		log.info("Running " + this.getClass().getName() + " plugin!");
 
+		Document doc;
+		try {
+			doc = Jsoup.parse(url, 10000);
+		} catch (IOException e1) {
+			throw new SystemErrorException();
+		}
+		
+		String caption = getCaption(doc);
+		log.info("caption: " + caption);
+		
+		String description = getDescription(doc);
+		log.info("description: " + description);
+		
+		String imageUrl = getImageUrl(doc);
+		log.info("image url: " + description);
+		
+		String urlString = null;
+		try {
+			urlString = getProperty(doc, "og:url");
+		} catch (Exception e) {
+			urlString = url.toString();
+		}
+
+		return DatastoreManager.getDatastore().createItem(caption, urlString, description, imageUrl);
+	}
+
+	/**
+	 * Get the caption.
+	 * @param doc Document
+	 * @return Caption
+	 */
+	protected String getCaption(Document doc) {
+
+		String caption = null;
+		try {
+			caption = getProperty(doc, "og:title");
+		} catch (Exception e) {}
+
+		if (caption == null) {
+			try {
+				caption = doc.select("title").first().text();
+			} catch (Exception e) {
+				caption = "";
+			}
+		}
+		
+		return caption;
+	}
+	
+	/**
+	 * Get the description.
+	 * @param doc Document
+	 * @return Description
+	 */
+	protected String getDescription(Document doc) {
+		
+		String description = null;
+		try {
+			description = getProperty(doc, "og:description");
+		} catch (Exception e) {}
+		
+		if (description == null) {
+			try {
+				description = doc.getElementsByAttributeValue("name", "description")
+				                 .first().attr("content");
+			} catch (Exception e) {
+				description = "";
+			}
+		}
+		
+		return description;
+	}
+
+	/**
+	 * Get the image url.
+	 * @param doc Document
+	 * @return Image url
+	 */
+	protected String getImageUrl(Document doc) {
+	
+		String imageUrl = "";
+		try {
+			imageUrl = getProperty(doc, "og:image");
+			imageUrl.replace(" ", "");
+			
+			// check if URL is a legit URL as og:url is broken on some sites 
+			// throws an exception if URL is not ok
+			URL u = new URL(imageUrl);
+			u.toURI(); 			
+		} catch (Exception e) {
+			imageUrl = "";
+		}
+		
+		return imageUrl;
+	}
+	
 	/**
 	 * Get the property content.
 	 * @param doc Document
