+++ src/de/kaffeeshare/server/plugins/Vimeo.java
@@ -15,15 +15,9 @@
  ******************************************************************************/
 package de.kaffeeshare.server.plugins;
 
-import java.io.IOException;
 import java.net.URL;
 
-import org.jsoup.Jsoup;
 import org.jsoup.nodes.Document;
-
-import de.kaffeeshare.server.datastore.DatastoreManager;
-import de.kaffeeshare.server.datastore.Item;
-import de.kaffeeshare.server.exception.SystemErrorException;
 
 /**
  * Plugin to handle vimeo pages.
@@ -37,75 +31,22 @@ public class Vimeo extends BasePlugin {
 	}
 
 	@Override
-	public Item createItem(URL url) {
-		log.info("Running Vimeo plugin!");
-
-		Document doc;
-		try {
-			doc = Jsoup.parse(url, 10000);
-		} catch (IOException e1) {
-			throw new SystemErrorException();
-		}
+	public String getDescription(Document doc) {
 
 		String videoId = null;
 		try {
 			videoId = getProperty(doc, "og:url").replace("http://vimeo.com/",
 					"");
-		} catch (Exception e) {
-		}
+		} catch (Exception e) {}
 
-		String caption = null;
-
-		try {
-			caption = getProperty(doc, "og:title");
-		} catch (Exception e) {
-		}
-
-		if (caption == null) {
-			try {
-				caption = doc.select("title").first().text();
-			} catch (Exception e) {
-				caption = "";
-			}
-		}
-
-		log.info("caption: " + caption);
-
-		String description = null;
-
-		try {
-			description = getProperty(doc, "og:description");
-		} catch (Exception e) {
-		}
-
-		if (description == null) {
-			try {
-				description = doc
-						.getElementsByAttributeValue("name", "description")
-						.first().attr("content");
-			} catch (Exception e) {
-				description = "";
-			}
-		}
-
+		String description = super.getDescription(doc);
 		if (videoId != null) {
 			description += "<br /><br /><br /><iframe src=\"http://player.vimeo.com/video/"
 					+ videoId
 					+ "?title=0&amp;byline=0&amp;portrait=0\" width=\"400\" height=\"225\" frameborder=\"0\" webkitAllowFullScreen mozallowfullscreen allowFullScreen></iframe>";
 		}
 
-		log.info("desc: " + description);
-
-		String imageUrl = "";
-
-		String urlString = null;
-		try {
-			urlString = getProperty(doc, "og:url");
-		} catch (Exception e) {
-			urlString = url.toString();
-		}
-
-		return DatastoreManager.getDatastore().createItem(caption,urlString, description, imageUrl);
+		return description;
 	}
 
 }
