+++ src/de/kaffeeshare/server/plugins/Youtube.java
@@ -15,17 +15,11 @@
  ******************************************************************************/
 package de.kaffeeshare.server.plugins;
 
-import java.io.IOException;
 import java.net.URL;
 
-import org.jsoup.Jsoup;
 import org.jsoup.nodes.Document;
 import org.jsoup.nodes.Element;
 import org.jsoup.select.Elements;
-
-import de.kaffeeshare.server.datastore.DatastoreManager;
-import de.kaffeeshare.server.datastore.Item;
-import de.kaffeeshare.server.exception.SystemErrorException;
 
 /**
  * Plugin to handle youtube pages.
@@ -39,15 +33,7 @@ public class Youtube extends BasePlugin {
 	}
 
 	@Override
-	public Item createItem(URL url) {
-		log.info("Running Youtube plugin!");
-
-		Document doc;
-		try {
-			doc = Jsoup.parse(url, 10000);
-		} catch (IOException e1) {
-			throw new SystemErrorException();
-		}
+	public String getDescription(Document doc) {
 
 		String videoId = null;
 		try {
@@ -62,58 +48,14 @@ public class Youtube extends BasePlugin {
 		} catch (Exception e) {
 		}
 
-		String caption = null;
-
-		try {
-			caption = getProperty(doc, "og:title");
-		} catch (Exception e) {
-		}
-
-		if (caption == null) {
-			try {
-				caption = doc.select("title").first().text();
-			} catch (Exception e) {
-				caption = "";
-			}
-		}
-
-		log.info("caption: " + caption);
-
-		String description = null;
-
-		try {
-			description = getProperty(doc, "og:description");
-		} catch (Exception e) {
-		}
-
-		if (description == null) {
-			try {
-				description = doc
-						.getElementsByAttributeValue("name", "description")
-						.first().attr("content");
-			} catch (Exception e) {
-				description = "";
-			}
-		}
-
+		String description = super.getDescription(doc);
 		if (videoId != null) {
 			description += "<br /><br /><br /><iframe width=\"560\" height=\"315\" src=\"http://www.youtube.com/embed/"
 					+ videoId
 					+ "\" frameborder=\"0\" allowfullscreen></iframe>";
 		}
 
-		log.info("desc: " + description);
-
-		String imageUrl = "";
-		
-		String urlString = null;
-		try {
-			urlString = getProperty(doc, "og:url");
-		} catch (Exception e) {
-			urlString = url.toString();
-		}
-
-		return DatastoreManager.getDatastore().createItem(caption,urlString, description, imageUrl);
+		return description;
 	}
 
 }
